pyproject.toml
[tool.poetry]
name = "poc-ing"
version = "0.1.0"
description = ""
authors = ["michaelsinanta <michaelchristlambert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.7.0"
python-dotenv = "^1.0.0"
gradio = "^4.14.0"
opencv-python = "^4.9.0.80"
streamlit = "^1.39.0"
pillow = "^10.4.0"
anthropic = "^0.34.2"
poppler-utils = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
